+++ engine/runtime-instrument-common/src/main/java/org/enso/interpreter/service/ExecutionService.java
@@ -77,6 +77,7 @@ import org.slf4j.LoggerFactory;
  * language.
  */
 public final class ExecutionService {
+  private static final Logger LOGGER = LoggerFactory.getLogger(ExecutionService.class);
   private static final String MAIN_METHOD = "main";
   private final EnsoContext context;
   private final Optional<IdExecutionService> idExecutionInstrument;
@@ -146,10 +147,9 @@ public final class ExecutionService {
     if (connectedLockManager != null) {
       connectedLockManager.connect(endpoint);
     } else {
-      LoggerFactory.getLogger(ExecutionService.class)
-          .warn(
-              "ConnectedLockManager was not initialized, even though a Language Server connection"
-                  + " has been established. This may result in synchronization errors.");
+      LOGGER.warn(
+          "ConnectedLockManager was not initialized, even though a Language Server connection"
+              + " has been established. This may result in synchronization errors.");
     }
   }
 
@@ -313,6 +313,7 @@ public final class ExecutionService {
    * @return a computation representing the evaluation of an expression
    */
   public CompletionStage<Object> evaluateExpression(Module module, String expression) {
+    LOGGER.trace("evaluateExpression in {} code: {}", module.getName(), expression);
     return submitExecution(() -> invoke.getCallTarget().call(module, expression));
   }